chore: move util to pkg
