+++ config/config.yaml
@@ -48,7 +48,7 @@ usagebackend:
   port: 443
 controller:
   host: controller
-  port: 8085
+  port: 3085
   https:
     cert:
     key:
